Added cli arg options for storage.tsdb.path etc. for podAggregator and federation. Moved MonitoringRole to crd/src/lib.
